fix: Adapt some tests for better handling of floating-point values

+++ SimpleCalc/Services/CalcImplementation.cs
@@ -64,12 +64,12 @@ namespace SimpleCalc.Services
             return firstValue / secondValue;
         }
 
-        public int Pow(int basis, int exponent)
+        public int Pow(double basis, double exponent)
         {
             return 0;
         }
 
-        public int Root(int value)
+        public int Root(double value)
         {
             return 0;
         }
